Integrated environment variables and refactored start page logic
