+++ website/src/components/RunButtonHandler/RunButtonHandler.js
@@ -5,7 +5,7 @@ export function handleRunButtonClick(selectedImage) {
 
 	const labels = ['0-2', '3-10', '11-17', '18-24', '25-40', '40-60', '60+'];
 
-	return fetch('http://localhost:8000/predict', {
+	return fetch(`${process.env.REACT_APP_API_URL}/predict`, {
 		method: 'POST',
 		headers: {
 			'Content-Type': 'application/json',
